rename luts
